+++ src/tg2go/bot/handlers/admin/commands/send.py
@@ -6,7 +6,7 @@ from aiogram.fsm.state import State, StatesGroup
 
 from tg2go.bot.lib.message.filters import AdminFilter
 from tg2go.bot.lib.message.io import ContextIO, SendMessage
-from tg2go.db.services.user_context import GetUserContextService
+from tg2go.db.services.user_context import GetContextService
 
 router = Router()
 
@@ -29,8 +29,8 @@ async def CommandSend(
         )
         return
 
-    ctx = await GetUserContextService()
-    chat_id = await ctx.GetTgChatIdBy(tg_username=command.args.replace("@", "").strip())
+    ctx = GetContextService()
+    chat_id = await ctx.GetChatIdByUsername(command.args.replace("@", "").strip())
 
     if chat_id is None:
         await SendMessage(