+++ src/tg2go/bot/handlers/admin/register.py
@@ -4,7 +4,6 @@ from tg2go.bot.handlers.admin.commands import (
     admin,
     blocking,
     logs,
-    messages,
     send,
     senda,
 )
@@ -14,7 +13,6 @@ def RegisterAdminHandlers(dp: Dispatcher) -> None:
     dp.include_routers(
         admin.router,
         logs.router,
-        messages.router,
         send.router,
         senda.router,
         blocking.router,