+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - DB Models & Repositories
 - Bot's library
+- Bot's admin commands
 
 ## [0.1.0] - 2025-06-15 - Vlad